Fix issue with removing child entities from weapons outside viewport.
Check to remove when going outside the viewport was verified on the update method of each entity, but the update method is never called if the element is not on viewport, so the element would still be added.
Setting alwaysUpdate to true allows for the update method to run outside of the viewport, where it will check that the element is not on the viewport, removing it from the game world.