Boss mode and player health.
Bosses fire weapons based on intervals, as defined on the enemy styles config.
Bosses spawn in the middle of the stage. Adapted enemy manager creation for this.
Laser entity with different behavior onCollision if spawned by an enemy entity.
Player no longer a sprite but an entity, so collisions can work.
Player now has a base health defined on game.data.